requirements.txt
Django==5.1.4
django-shopping-cart==0.1
asgiref==3.8.1
Pillow==11.0.0
pip==24.3.1
sqlparse==0.5.3
tzdata==2024.2